+++ docs/release_notes.md
@@ -1,5 +1,17 @@
 # Release Notes
 
+## Version 1.1.0 (May 18, 2023)
+
+* CLI updates for token-service
+* CLI updates for streaming-service
+* Validation added for tasking requests
+* Updated documentation for CLI commands, discovery-service, monitoring-service, ordering-service, and tasking-service
+* Added Jupyter notebook examples for discovery-service, ordering-service, and tasking-service
+* Added get_stac_item functionality to discovery-service
+* Added get_all_pipelines functionality to ordering-service
+* Removed GDAL and rasterio as dependencies for package
+* Minor bug fixes
+
 ## Version 1.0.0 (May 10, 2022)
 
 * Initial release